Add HomeComponent and update environment settings
A new HomeComponent has been added, including its corresponding HTML file and spec file for testing. The RouterModule has been altered to navigate to the HomeComponent which signifies a change in the project architecture to accommodate the new component.
The environment variable `showCase` has been added to both `environment.ts` and `environment.prod.ts` and used in `app.layout.service.ts`. This is to control some layout features based on the environment without changing the code when deploying the app. The README.md file was corrected.