Update HomeComponent and package dependencies
HomeComponent has been updated with a constructor and a console.log for creation event. Dependencies in package.json and package-lock.json have been updated for betterment of the project. This includes an update in version of @angular-devkit/build-angular and @typescript-eslint packages among others. This commit ensures application uses latest dependencies to avoid any possible security issues and provides the latest features and optimizations.